% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{CalcConcordance}
\alias{CalcConcordance}
\alias{CalcConcordance,}
\alias{CalcConcordanceDiscordance}
\alias{CalcDiscordance,}
\title{CalcConcordance}
\usage{
CalcConcordance(logitMod)
}
\arguments{
\item{logitMod}{A logit model}
}
\value{
a list containing percentage of concordant pairs, percentage discordant pairs, percentage ties and No. of pairs.
}
\description{
Calculate concordance and discordance percentages for a logit model
}
\details{
Calculate the percentage of concordant and discordant pairs for a given logit model.
}
\examples{
accept <- c (1, 0, 1, 0, 1, 1, 0, 0, 0,1, 0, 1, 0, 0, 1)
acad   <- c (66, 60, 80, 60, 52, 60, 47, 90, 75, 35, 46, 75, 66, 54, 76)
sports <- c (2.6,4.6,4.5, 3.3, 3.13, 4, 1.9, 3.5, 1.2, 1.8, 1, 5.1, 3.3, 5.2, 4.9)
rank   <- c (3, 3, 1, 4, 4, 2, 4, 4, 4, 3, 3, 3, 2, 2, 1)
inputData  <- data.frame (accept, acad , sports, rank) # assemble the data frame
logitModel <- glm(accept ~ ., family="binomial", data = inputData )
CalcConcordance(logitModel)
}
\author{
Selva Prabhakaran
}

