% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{getWOETable}
\alias{getWOETable}
\title{getWOETable}
\usage{
getWOETable(X = X, Y = Y, valueOfGood = 1)
}
\arguments{
\item{X}{A categorical variable for which WOE is to be calculated.}

\item{Y}{A binary response variable}

\item{valueOfGood}{The value of denoting the occurence of the event of interest in the response (Y) variable.}
}
\value{
The WOE table containing the WOE and information values for each class.
}
\description{
Compute the woe table that contains the weight of evidence for respective classes in a categorical variable.
}
\details{
For a given binary response variable (Y) and a categorical (X) variable, compute the WOE and IVs for respective classes.
}
\examples{
data("OJ")
X <- factor(as.character(OJ$STORE))  # prepares X
Y <- as.character(OJ$Purchase)
valueOfGood <- "CH"  # input argument
getWOETable(X, Y, valueOfGood)
}
\author{
Selva Prabhakaran
}

