% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{iv}
\alias{iv}
\title{iv}
\usage{
iv(X = X, Y = Y, valueOfGood = 1)
}
\arguments{
\item{X}{A categorical variable for which WOE is to be calculated.}

\item{Y}{A binary response variable}

\item{valueOfGood}{The value of denoting the occurence of the event of interest in the response (Y) variable.}
}
\value{
The IV for a categorical variable of interest in predicting a binary response variable.
}
\description{
Compute the information value of a predictor categorical variable to explain a response variable.
}
\details{
Compute the information value for a categorical variable. An IV of greater than 0.1 implies a high predictive power while that below 0.03 shows its not predictive.
}
\examples{
data("OJ")
X <- factor(as.character(OJ$STORE))  # prepares X
Y <- as.character(OJ$Purchase)
valueOfGood <- "CH"  # input argument
getIV(X, Y, valueOfGood)
}
\author{
Selva Prabhakaran
}

