% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{kappaCohen}
\alias{kappaCohen}
\title{kappaCohen}
\usage{
kappaCohen(logitMod, threshold = 0.5)
}
\arguments{
\item{logitMod}{A logit model}

\item{threshold}{If predicted value is above the threshold, it will be considered as an event (1), else it will be a non-event (0). Defaults to 0.5.}
}
\value{
The youdensIndex of the logit model, which is calculated as Sensitivity + Specificity - 1
}
\description{
Calculate the Cohen's kappa statistic for a given logit model.
}
\details{
For a given logit model, Youden's index is calculated as sensitivity + specificity - 1
}
\examples{
accept <- c (1, 0, 1, 0, 1, 1, 0, 0, 0,1, 0, 1, 0, 0, 1)
acad   <- c (66, 60, 80, 60, 52, 60, 47, 90, 75, 35, 46, 75, 66, 54, 76)
sports <- c (2.6,4.6,4.5, 3.3, 3.13, 4, 1.9, 3.5, 1.2, 1.8, 1, 5.1, 3.3, 5.2, 4.9)
rank   <- c (3, 3, 1, 4, 4, 2, 4, 4, 4, 3, 3, 3, 2, 2, 1)
inputData  <- data.frame (accept, acad , sports, rank) # assemble the data frame
logitModel <- glm(accept ~ ., family="binomial", data = inputData )
kappaCohen(logitMod=logitModel)
}
\author{
Selva Prabhakaran
}

