% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{misClassError}
\alias{misClassError}
\title{misClassError}
\usage{
misClassError(logitMod, threshold = 0.5)
}
\arguments{
\item{logitMod}{A logit model}

\item{threshold}{If predicted value is above the threshold, it will be considered as an event (1), else it will be a non-event (0).}
}
\value{
The misclassification error, which tells what proportion of predicted direction did not match with the actuals.
}
\description{
Calculate the percentage misclassification error for this logit model's fitted values.
}
\details{
For a given logit model, misclassfication error is the number of mismatches between the predicted and actuals direction of the binary y variable.
}
\examples{
accept <- c (1, 0, 1, 0, 1, 1, 0, 0, 0,1, 0, 1, 0, 0, 1)
acad   <- c (66, 60, 80, 60, 52, 60, 47, 90, 75, 35, 46, 75, 66, 54, 76)
sports <- c (2.6,4.6,4.5, 3.3, 3.13, 4, 1.9, 3.5, 1.2, 1.8, 1, 5.1, 3.3, 5.2, 4.9)
rank   <- c (3, 3, 1, 4, 4, 2, 4, 4, 4, 3, 3, 3, 2, 2, 1)
inputData  <- data.frame (accept, acad , sports, rank) # assemble the data frame
logitModel <- glm(accept ~ ., family="binomial", data = inputData )
misClassError(logitMod=logitModel)
}
\author{
Selva Prabhakaran
}

