% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Main.R
\name{plotROC}
\alias{plotROC}
\title{plotROC}
\usage{
plotROC(actuals, predictedScores, threshold = 0.5, Show.labels = F)
}
\arguments{
\item{actuals}{The actual binary flags for the response variable. It can take values of either 1 or 0, where 1 represents the 'Good' or 'Events' while 0 represents 'Bad' or 'Non-Events'.}

\item{predictedScores}{The prediction probability scores for each observation.}

\item{threshold}{If predicted value is above the threshold, it will be considered as an event (1), else it will be a non-event (0). Defaults to 0.5.}

\item{Show.labels}{Whether the probability scores should be printed at change points?. Defaults to False.}
}
\value{
Plots the ROC curve
}
\description{
Plot the Receiver Operating Characteristics(ROC) Curve based on ggplot2
}
\details{
For a given actuals and predicted probability scores, A ROC curve is plotted using the ggplot2 framework along the the area under the curve.
}
\examples{
data('ActualsAndScores')
plotROC(actuals=ActualsAndScores$Actuals, predictedScores=ActualsAndScores$PredictedScores)
}
\author{
Selva Prabhakaran
}

