% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{SomersD}
\alias{SomersD}
\alias{SomersDistance}
\title{SomersD}
\usage{
SomersD(logitMod)
}
\arguments{
\item{logitMod}{A logit model}
}
\value{
The Somers D statistic, which tells how many more concordant than discordant pairs exist divided by total number of pairs.
}
\description{
Calculate the Somers D statistic for a given logit model
}
\details{
For a given logit model, Somer's D is calculated as the number of concordant pairs less number of discordant pairs divided by total number of pairs.
}
\examples{
accept <- c (1, 0, 1, 0, 1, 1, 0, 0, 0,1, 0, 1, 0, 0, 1)
acad   <- c (66, 60, 80, 60, 52, 60, 47, 90, 75, 35, 46, 75, 66, 54, 76)
sports <- c (2.6,4.6,4.5, 3.3, 3.13, 4, 1.9, 3.5, 1.2, 1.8, 1, 5.1, 3.3, 5.2, 4.9)
rank   <- c (3, 3, 1, 4, 4, 2, 4, 4, 4, 3, 3, 3, 2, 2, 1)
inputData  <- data.frame (accept, acad , sports, rank) # assemble the data frame
logitModel <- glm(accept ~ ., family="binomial", data = inputData )
SomersD(logitMod=logitModel)
}
\author{
Selva Prabhakaran
}

