% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{woe}
\alias{woe}
\title{woe}
\usage{
woe(X = X, Y = Y, valueOfGood = 1)
}
\arguments{
\item{X}{A categorical variable for which WOE is to be calculated.}

\item{Y}{A binary response variable}

\item{valueOfGood}{The value of denoting the occurence of the event of interest in the response (Y) variable.}
}
\value{
The WOE for the categorical variable of interest.
}
\description{
Compute the weights of evidence for a categorical variable.
}
\details{
Compute the weights of evidence for a categorical variable so it can be used as predictor in place of the categorical variable itself.
}
\examples{
data("OJ")
X <- factor(as.character(OJ$STORE))  # prepares X
Y <- as.character(OJ$Purchase)
valueOfGood <- "CH"  # input argument
getWOE(X, Y, valueOfGood)
}
\author{
Selva Prabhakaran
}

