% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{plotROC}
\alias{plotROC}
\title{plotROC}
\usage{
plotROC(logitMod)
}
\arguments{
\item{logitMod}{A logit model}
}
\value{
Plots the ROC curve
}
\description{
Plot the Receiver Operating Characteristics(ROC) Curve based on ggplot2
}
\details{
For a given logit model, A ROC curve is plotted using the ggplot2 framework along the the area under the curve.
}
\examples{
accept <- c (1, 0, 1, 0, 1, 1, 0, 0, 0,1, 0, 1, 0, 0, 1)
acad   <- c (66, 60, 80, 60, 52, 60, 47, 90, 75, 35, 46, 75, 66, 54, 76)
sports <- c (2.6,4.6,4.5, 3.3, 3.13, 4, 1.9, 3.5, 1.2, 1.8, 1, 5.1, 3.3, 5.2, 4.9)
rank   <- c (3, 3, 1, 4, 4, 2, 4, 4, 4, 3, 3, 3, 2, 2, 1)
inputData  <- data.frame (accept, acad , sports, rank) # assemble the data frame
logitModel <- glm(accept ~ ., family="binomial", data = inputData )
plotROC(logitMod=logitModel)
}
\author{
Selva Prabhakaran
}

